Fix TicksEveryEightMeasures calculation becoming 0 with denominators larger than 4 (integer division skill issue)

+++ YARG.Core/Engine/BaseEngine.cs
@@ -220,7 +220,7 @@ namespace YARG.Core.Engine
 
             var currentTimeSig = timeSigs[State.CurrentTimeSigIndex];
 
-            State.TicksEveryEightMeasures = Resolution * (4 / currentTimeSig.Denominator) * currentTimeSig.Numerator * 8;
+            State.TicksEveryEightMeasures = (uint)(Resolution * ((double)4 / currentTimeSig.Denominator) * currentTimeSig.Numerator * 8);
         }
 
         public override void Reset(bool keepCurrentButtons = false)
